tracers moved to their own files, cleaned up trace imports
