+++ ethcore/src/common.rs
@@ -26,4 +26,3 @@ pub use transaction::*;
 pub use log_entry::*;
 pub use receipt::*;
 pub use action_params::*;
-pub use trace::*;
