+++ ethcore/src/trace/mod.rs
@@ -16,14 +16,16 @@
 
 //! Tracing
 
-mod trace;
+pub mod trace;
 mod filter;
+mod noop_tracer;
+mod executive_tracer;
 
-pub use self::trace::*;
+pub use self::trace::{Trace, TraceCall, TraceCreate, TraceAction};
 pub use self::filter::Filter;
-use util::bytes::Bytes;
-use util::hash::Address;
-use util::numbers::U256;
+pub use self::noop_tracer::NoopTracer;
+pub use self::executive_tracer::ExecutiveTracer;
+use util::{Bytes, Address, U256};
 use action_params::ActionParams;
 
 /// This trait is used by executive to build traces.
@@ -70,134 +72,4 @@ pub trait Tracer: Send {
 
 	/// Consumes self and returns all traces.
 	fn traces(self) -> Vec<Trace>;
-}
-
-/// Nonoperative tracer. Does not trace anything.
-pub struct NoopTracer;
-
-impl Tracer for NoopTracer {
-	fn prepare_trace_call(&self, _: &ActionParams) -> Option<TraceCall> {
-		None
-	}
-
-	fn prepare_trace_create(&self, _: &ActionParams) -> Option<TraceCreate> {
-		None
-	}
-
-	fn prepare_trace_output(&self) -> Option<Bytes> {
-		None
-	}
-
-	fn trace_call(&mut self, call: Option<TraceCall>, _: U256, output: Option<Bytes>, _: usize, _: Vec<Trace>,
-				  _: bool) {
-		assert!(call.is_none());
-		assert!(output.is_none());
-	}
-
-	fn trace_create(&mut self, create: Option<TraceCreate>, _: U256, code: Option<Bytes>, _: Address, _: usize, _: Vec<Trace>) {
-		assert!(create.is_none());
-		assert!(code.is_none());
-	}
-
-	fn trace_failed_call(&mut self, call: Option<TraceCall>, _: usize, _: Vec<Trace>, _: bool) {
-		assert!(call.is_none());
-	}
-
-	fn trace_failed_create(&mut self, create: Option<TraceCreate>, _: usize, _: Vec<Trace>) {
-		assert!(create.is_none());
-	}
-
-	fn subtracer(&self) -> Self {
-		NoopTracer
-	}
-
-	fn traces(self) -> Vec<Trace> {
-		vec![]
-	}
-}
-
-/// Simple executive tracer. Traces all calls and creates. Ignores delegatecalls.
-#[derive(Default)]
-pub struct ExecutiveTracer {
-	traces: Vec<Trace>
-}
-
-impl Tracer for ExecutiveTracer {
-	fn prepare_trace_call(&self, params: &ActionParams) -> Option<TraceCall> {
-		Some(TraceCall::from(params.clone()))
-	}
-
-	fn prepare_trace_create(&self, params: &ActionParams) -> Option<TraceCreate> {
-		Some(TraceCreate::from(params.clone()))
-	}
-
-	fn prepare_trace_output(&self) -> Option<Bytes> {
-		Some(vec![])
-	}
-
-	fn trace_call(&mut self, call: Option<TraceCall>, gas_used: U256, output: Option<Bytes>, depth: usize, subs:
-				  Vec<Trace>, delegate_call: bool) {
-		// don't trace if it's DELEGATECALL or CALLCODE.
-		if delegate_call {
-			return;
-		}
-
-		let trace = Trace {
-			depth: depth,
-			subs: subs,
-			action: TraceAction::Call(call.expect("Trace call expected to be Some.")),
-			result: TraceResult::Call(TraceCallResult {
-				gas_used: gas_used,
-				output: output.expect("Trace call output expected to be Some.")
-			})
-		};
-		self.traces.push(trace);
-	}
-
-	fn trace_create(&mut self, create: Option<TraceCreate>, gas_used: U256, code: Option<Bytes>, address: Address, depth: usize, subs: Vec<Trace>) {
-		let trace = Trace {
-			depth: depth,
-			subs: subs,
-			action: TraceAction::Create(create.expect("Trace create expected to be Some.")),
-			result: TraceResult::Create(TraceCreateResult {
-				gas_used: gas_used,
-				code: code.expect("Trace create code expected to be Some."),
-				address: address
-			})
-		};
-		self.traces.push(trace);
-	}
-
-	fn trace_failed_call(&mut self, call: Option<TraceCall>, depth: usize, subs: Vec<Trace>, delegate_call: bool) {
-		// don't trace if it's DELEGATECALL or CALLCODE.
-		if delegate_call {
-			return;
-		}
-
-		let trace = Trace {
-			depth: depth,
-			subs: subs,
-			action: TraceAction::Call(call.expect("Trace call expected to be Some.")),
-			result: TraceResult::FailedCall,
-		};
-		self.traces.push(trace);
-	}
-
-	fn trace_failed_create(&mut self, create: Option<TraceCreate>, depth: usize, subs: Vec<Trace>) {
-		let trace = Trace {
-			depth: depth,
-			subs: subs,
-			action: TraceAction::Create(create.expect("Trace create expected to be Some.")),
-			result: TraceResult::FailedCreate,
-		};
-		self.traces.push(trace);
-	}
-
-	fn subtracer(&self) -> Self {
-		ExecutiveTracer::default()
-	}
-
-	fn traces(self) -> Vec<Trace> {
-		self.traces
-	}
 }
